requirements.txt
django
djangorestframework
pandas
sqlalchemy
psycopg2-binary

